Добавить заказы в хранилище заказов 2

+++ Store.Web/SessionExtensions.cs
@@ -32,7 +32,7 @@ namespace Store.Web
                 {
                     var orderId=reader.ReadInt32();
                     var totalCount = reader.ReadInt32();
-                    var totalPrice = reader.ReadInt32();
+                    var totalPrice = reader.ReadDecimal();
 
                     cart = new Cart(orderId)
                     {
